pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "procurement-system"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.15",
    "django-filter>=24.0",
    "django-cors-headers>=4.0",
    "msal>=1.28",
    "gunicorn>=22.0",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "click>=8.1",
    "rich>=13.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=5.0",
    "factory-boy>=3.3",
    "ruff>=0.5",
]

[project.scripts]
procurement-cli = "cli.main:cli"

[tool.setuptools.packages.find]
include = ["config*", "accounts*", "core*", "orders*", "payments*", "deliveries*", "approvals*", "assets*", "cli*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["tests.py", "test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--tb=short -q"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
